% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMinRT.R
\name{computeMinRT}
\alias{computeMinRT}
\title{Compute minimum Reaction Time}
\usage{
computeMinRT(RT, correct, bins = seq(min(RT, na.rm = T), max(RT, na.rm =
  T) + 10, by = 10), nBinSucc = 5, method = "chi2", thresh = 0.05)
}
\arguments{
\item{RT}{Reaction times vector (in ms or adapt bins)}

\item{correct}{Correctness vector (coerced to logical)}

\item{nBinSucc}{Number of successive significantly above chance bins}

\item{method}{chi square. (unused at present)}

\item{thresh}{threshold for significance of the chi square test.}
}
\value{
bins Bin limits for computation

minRT the minimum Reaction Time
}
\description{
Compute minimum RT above which subjects are above chance in a 2AFC task, given RT and correctness vectors
}
