% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{dpc2perf}
\alias{dpc2perf}
\title{performance predictor based on dprime (dp) and criterion (c)}
\usage{
dpc2perf(dp, c)
}
\arguments{
\item{dp}{d'}

\item{c}{criterion}
}
\value{
a list with Hit and False Alarm rates
}
\description{
Predict Hit Rate and False Alarm Rates based on dprime (dp) and criterion (c) in a Yes/No task
}

