% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outliers.R
\name{find.outliers}
\alias{find.outliers}
\title{Find outliers}
\usage{
find.outliers(x, tail = "both", meth = "trim", probs = c(0.25, 0.75),
  na.rm = TRUE, ...)
}
\arguments{
\item{x}{A vector}

\item{probs}{A 1 or 2 element numeric vector of probabilities}

\item{na.rm}{A boolean whether to remove NA}
}
\value{
A cleaned vector where all outliers outside of \code{quantile(x,probs)} have been removed.
}
\description{
Return logical true where outliers are in a vector.
}
\examples{
x <- randn(1000)
y <- find.outliers(x,probs=c(.01,.99))
}

