% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtextinfiles.R
\name{findtextinfiles}
\alias{findtextinfiles}
\title{Find text in files}
\usage{
findtextinfiles(tofind, filter = "\\\\.R(md)?$", dir = getwd(),
  full.names = T, recursive = F)
}
\arguments{
\item{tofind}{string: the pattern to find in all files}

\item{filter}{string: the pattern of file names to search in (default to \code{\.R(md)?$} = files ending in .R or .Rmd)}

\item{dir}{the directory root to search in (default to \code{getwd()})}

\item{full.names}{logical: Whether to return full path to files or not (default \code{TRUE})}

\item{recursive}{logical: Whether to search subdirectories (default \code{FALSE})}
}
\description{
Print file names and line numbers where a text pattern is found in all text files within in a specified directory (and subdirectories)
}
\examples{
findtextinfiles('function')

}

