\name{MaxPac-package}
\alias{MaxPac-package}
\alias{MaxPac}
\docType{package}
\title{
MaxPac
}
\description{
These are my misc functions. Handy tools for my data analysis.
}
\details{
\tabular{ll}{
Package: \tab MaxPac\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-18\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Maximilien Chaumon
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
