% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{perf2Ap}
\alias{perf2Ap}
\title{YN Aprime estimator}
\usage{
perf2Ap(hitR, faR)
}
\arguments{
\item{hitR}{Hit Rate}

\item{faR}{False Alarm Rate}
}
\value{
Aprime value
}
\description{
Estimate Aprime from Hit Rate and False Alarm Rate in a Yes/No task
Formula (3) in Stanislaw and Todorov 1999
}
