% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{perf2B}
\alias{perf2B}
\title{YN bias (B) estimator}
\usage{
perf2B(hitR, faR)
}
\arguments{
\item{hitR}{Hit Rate}

\item{faR}{False Alarm Rate}
}
\value{
B value
}
\description{
Estimate bias (B) from Hit Rate and False Alarm Rate in a Yes/No task
Formula (5) in Stanislaw and Todorov 1999 (note typo in formula, but not in software implementations later in pdf)
}
