% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{perf2Bsecond}
\alias{perf2Bsecond}
\title{YN B" estimator}
\usage{
perf2Bsecond(hitR, faR)
}
\arguments{
\item{hitR}{Hit Rate}

\item{faR}{False Alarm Rate}
}
\value{
B" value
}
\description{
Estimate B" from Hit Rate and False Alarm Rate in a Yes/No task
Formula (9) in Stanislaw and Todorov 1999
}
