% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{perf2c}
\alias{perf2c}
\title{YN criterion (c) estimator}
\usage{
perf2c(hitR, faR)
}
\arguments{
\item{hitR}{Hit Rate}

\item{faR}{False Alarm Rate}
}
\value{
dprime value
}
\description{
Estimate criterion (c) from Hit Rate and False Alarm Rate in a Yes/No task
}

