% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{perf2dp}
\alias{perf2dp}
\title{YN dprime estimator}
\usage{
perf2dp(hitR, faR)
}
\arguments{
\item{hitR}{Hit Rate}

\item{faR}{False Alarm Rate}
}
\value{
dprime value
}
\description{
Estimate dprime from Hit Rate and False Alarm Rate in a Yes/No task
}

