% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_helper.R
\name{perf2dpc}
\alias{perf2dpc}
\title{YN dprime (dp) and criterion (c) estimator}
\usage{
perf2dpc(hitR, faR)
}
\arguments{
\item{hitR}{Hit Rate}

\item{faR}{False Alarm Rate}
}
\value{
a list with dp and c values
}
\description{
Estimate criterion (c) from Hit Rate and False Alarm Rate in a Yes/No task
}

