% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{rm.outliers}
\alias{rm.outliers}
\title{Remove outliers}
\usage{
rm.outliers(x, tail = "both", meth = "trim", probs = c(0.25, 0.75),
  sd = 2, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A vector}

\item{tail}{One of \code{'both'}, \code{'left'}, or \code{'right'}}

\item{meth}{One of \code{'trim'} or \code{'IQR'}}

\item{probs}{A 1 or 2 element numeric vector of probabilities}

\item{na.rm}{A boolean whether to remove NA}

\item{sdn}{when meth = 'sd', number of standard deviations away from the mean to detect outlier}
}
\value{
A cleaned vector where all outliers outside of \code{quantile(x,probs)} (see details for individual methods) are changed to NA.
}
\description{
Remove outliers from a vector.
}
\details{
Each method will be described here in the future.
}
\examples{
x <- rnorm(1000)
y <- rm.outliers(x,probs=c(.01,.99))

}

