% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd.trim}
\alias{sd.trim}
\title{trimmed standard deviation}
\usage{
sd.trim(x, na.rm = FALSE, trim = 0, ...)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{A boolean}

\item{trim}{Numeric where to trim (proportion)}
}
\description{
compute standard deviation after trimming...
}

