% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{find.outliers}
\alias{find.outliers}
\title{Find outliers}
\usage{
find.outliers(x, tail = "both", meth = "trim", probs = c(0.25, 0.75),
  sdn = 2, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A vector}

\item{tail}{One of \code{'both'}, \code{'left'}, or \code{'right'}}

\item{meth}{One of \code{'trim'}, \code{'sd'} or \code{'IQR'}}

\item{probs}{A 1 or 2 element numeric vector of probabilities}

\item{sdn}{when meth = 'sd', number of standard deviations away from the mean to detect outlier}

\item{na.rm}{A boolean whether to remove NA}
}
\value{
A boolean vector where outliers outside of \code{quantile(x,probs)} (see details for individual methods) are true.
}
\description{
Return logical true where outliers are in a vector.
}
\details{
Each method will be described here in the future.
}
\examples{
x <- rnorm(1000)
y <- find.outliers(x,probs=c(.01,.99))

}

