\name{MaxPac-package}
\alias{MaxPac-package}
\alias{MaxPac}
\docType{package}
\title{
MaxPac
}
\description{
These are my misc functions. Handy tools for my data analysis.
}
\details{
\tabular{ll}{
Package: \tab MaxPac\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-18\cr
License: \tab What license is it under?\cr
}
}
\author{
Maximilien Chaumon
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
