% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outliers.R
\name{rm.outliers}
\alias{rm.outliers}
\title{Remove outliers}
\usage{
rm.outliers(x, tail = "both", meth = "trim", probs = c(0.25, 0.75),
  na.rm = TRUE, ...)
}
\arguments{
\item{x}{A vector}

\item{tail}{One of \code{'both'}, \code{'left'}, or \code{'right'}}

\item{meth}{One of \code{'trim'} or \code{'IQR'}}

\item{probs}{A 1 or 2 element numeric vector of probabilities}

\item{na.rm}{A boolean whether to remove NA}
}
\value{
A cleaned vector where all outliers outside of \code{quantile(x,probs)} have been removed.
}
\description{
Remove outliers from a vector.
}
\examples{
x <- randn(1000)
y <- rm.outliers(x,probs=c(.01,.99))
}

